--- forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/api/LayeredTexture.java
@@ -1,5 +1,0 @@
-package com.sigmundgranaas.forgero.model.api;
-
-import java.util.List;
-
-public record LayeredTexture(List<TextureLayer> layers) {}
