--- forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/api/ModelProvider.java
@@ -1,6 +1,0 @@
-package com.sigmundgranaas.forgero.model.api;
-
-import com.sigmundgranaas.forgero.common.identifier.api.OpenIdentifier;
-import java.util.Optional;
-
-public interface ModelProvider { Optional<Model> get(OpenIdentifier id); }
