--- forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/api/ModelRegistrationService.java
@@ -1,3 +1,0 @@
-package com.sigmundgranaas.forgero.model.api;
-
-public interface ModelRegistrationService { void registerModels(String namespace); }
