--- forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/api/ModelRegistry.java
@@ -1,10 +1,0 @@
-package com.sigmundgranaas.forgero.model.api;
-
-import com.sigmundgranaas.forgero.common.identifier.api.OpenIdentifier;
-
-import java.util.Optional;
-
-public interface ModelRegistry {
-	Optional<Model> find(OpenIdentifier id);
-	void register(Model model);
-}
