--- forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/api/ModelResolver.java
@@ -1,7 +1,0 @@
-package com.sigmundgranaas.forgero.model.api;
-
-import com.sigmundgranaas.forgero.core.component.api.Component;
-
-import java.util.Optional;
-
-public interface ModelResolver { Optional<LayeredTexture> resolve(Component component); }
