--- forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/api/TextureCompositor.java
@@ -1,5 +1,0 @@
-package com.sigmundgranaas.forgero.model.api;
-
-import java.awt.image.BufferedImage;
-
-public interface TextureCompositor { BufferedImage render(LayeredTexture model); }
