--- forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/api/TextureLayer.java
@@ -1,7 +1,0 @@
-package com.sigmundgranaas.forgero.model.api;
-
-import java.util.Comparator;
-
-public record TextureLayer(String texture, int order) implements Comparable<TextureLayer> {
-	@Override public int compareTo(TextureLayer other) { return Comparator.comparingInt(TextureLayer::order).compare(this, other); }
-}
