--- forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/api/TextureProvider.java
@@ -1,6 +1,0 @@
-package com.sigmundgranaas.forgero.model.api;
-
-import java.awt.image.BufferedImage;
-import java.util.Optional;
-
-public interface TextureProvider { Optional<BufferedImage> getTexture(String identifier); }
