--- forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/impl/dto/ModelDTO.java
@@ -1,5 +1,0 @@
-package com.sigmundgranaas.forgero.model.impl.dto;
-
-public interface ModelDTO {
-	String model_type();
-}
