--- forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/impl/dto/ModelSelectorDTO.java
@@ -1,5 +1,0 @@
-package com.sigmundgranaas.forgero.model.impl.dto;
-
-import org.jetbrains.annotations.Nullable;
-
-public record ModelSelectorDTO(String target_slot, @Nullable String target_tag, String model) {}
