--- forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/impl/dto/PartDTO.java
@@ -1,3 +1,0 @@
-package com.sigmundgranaas.forgero.model.impl.dto;
-
-public record PartDTO(String slot, int order) {}
