--- forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/impl/dto/SelfDTO.java
@@ -1,3 +1,0 @@
-package com.sigmundgranaas.forgero.model.impl.dto;
-
-public record SelfDTO(String texture, int order) {}
