--- forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/impl/dto/StaticModelDTO.java
@@ -1,10 +1,0 @@
-package com.sigmundgranaas.forgero.model.impl.dto;
-
-import com.google.gson.JsonElement;
-import org.jetbrains.annotations.Nullable;
-
-public record StaticModelDTO(
-		String model_type,
-		String texture,
-		@Nullable JsonElement display
-) implements ModelDTO {}
