+++ forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/loading/api/ModelProvider.java
@@ -1,0 +1,8 @@
+package com.sigmundgranaas.forgero.model.loading.api;
+
+import com.sigmundgranaas.forgero.common.identifier.api.OpenIdentifier;
+import com.sigmundgranaas.forgero.model.api.Model;
+
+import java.util.Optional;
+
+public interface ModelProvider { Optional<Model> get(OpenIdentifier id); }
