+++ forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/loading/impl/dto/ModelDTO.java
@@ -1,0 +1,5 @@
+package com.sigmundgranaas.forgero.model.loading.impl.dto;
+
+public interface ModelDTO {
+	String model_type();
+}
