+++ forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/loading/impl/dto/ModelSelectorDTO.java
@@ -1,0 +1,5 @@
+package com.sigmundgranaas.forgero.model.loading.impl.dto;
+
+import org.jetbrains.annotations.Nullable;
+
+public record ModelSelectorDTO(String target_slot, @Nullable String target_tag, String model) {}
