+++ forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/loading/impl/dto/PartDTO.java
@@ -1,0 +1,3 @@
+package com.sigmundgranaas.forgero.model.loading.impl.dto;
+
+public record PartDTO(String slot, int order) {}
