+++ forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/loading/impl/dto/SelfDTO.java
@@ -1,0 +1,3 @@
+package com.sigmundgranaas.forgero.model.loading.impl.dto;
+
+public record SelfDTO(String texture, int order) {}
