+++ forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/registry/api/ModelRegistrationService.java
@@ -1,0 +1,3 @@
+package com.sigmundgranaas.forgero.model.registry.api;
+
+public interface ModelRegistrationService { void registerModels(String namespace); }
