+++ forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/rendering/api/TextureCompositor.java
@@ -1,0 +1,7 @@
+package com.sigmundgranaas.forgero.model.rendering.api;
+
+import com.sigmundgranaas.forgero.model.resolution.api.LayeredTexture;
+
+import java.awt.image.BufferedImage;
+
+public interface TextureCompositor { BufferedImage render(LayeredTexture model); }
