+++ forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/rendering/api/TextureProvider.java
@@ -1,0 +1,6 @@
+package com.sigmundgranaas.forgero.model.rendering.api;
+
+import java.awt.image.BufferedImage;
+import java.util.Optional;
+
+public interface TextureProvider { Optional<BufferedImage> getTexture(String identifier); }
