+++ forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/rendering/impl/PngWriter.java
@@ -1,0 +1,15 @@
+package com.sigmundgranaas.forgero.model.rendering.impl;
+
+import javax.imageio.ImageIO;
+
+import java.awt.image.BufferedImage;
+import java.io.File;
+import java.io.IOException;
+
+public class PngWriter {
+	public static void save(BufferedImage image, String path) throws IOException {
+		File outputFile = new File(path);
+		outputFile.getParentFile().mkdirs();
+		ImageIO.write(image, "PNG", outputFile);
+	}
+}
