+++ forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/resolution/api/LayeredTexture.java
@@ -1,0 +1,5 @@
+package com.sigmundgranaas.forgero.model.resolution.api;
+
+import java.util.List;
+
+public record LayeredTexture(List<TextureLayer> layers) {}
