+++ forgero-core-2/src/main/java/com/sigmundgranaas/forgero/model/resolution/api/ModelResolver.java
@@ -1,0 +1,7 @@
+package com.sigmundgranaas.forgero.model.resolution.api;
+
+import com.sigmundgranaas.forgero.core.component.api.Component;
+
+import java.util.Optional;
+
+public interface ModelResolver { Optional<LayeredTexture> resolve(Component component); }
