+++ forgero-core-2/src/test/java/com/sigmundgranaas/forgero/model/ModelRegistrationServiceTest.java
@@ -1,0 +1,26 @@
+package com.sigmundgranaas.forgero.model;
+
+import com.sigmundgranaas.forgero.common.identifier.api.OpenIdentifier;
+
+import com.sigmundgranaas.forgero.model.registry.api.ModelRegistrationService;
+import com.sigmundgranaas.forgero.model.registry.api.ModelRegistry;
+import com.sigmundgranaas.forgero.model.registry.impl.DefaultModelRegistrationService;
+import com.sigmundgranaas.forgero.model.registry.impl.MapBackedModelRegistry;
+import com.sigmundgranaas.forgero.utility.resource.loader.implementation.ClassPathResourceProvider;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+public class ModelRegistrationServiceTest {
+	@Test
+	void testRegisterModels() {
+		ModelRegistry registry = new MapBackedModelRegistry();
+		ModelRegistrationService service = new DefaultModelRegistrationService(registry, new ClassPathResourceProvider("assets"));
+		service.registerModels("forgero");
+
+		assertTrue(registry.find(new OpenIdentifier("forgero:oak-handle")).isPresent());
+		assertTrue(registry.find(new OpenIdentifier("forgero:iron-pickaxe_head")).isPresent());
+		assertTrue(registry.find(new OpenIdentifier("forgero:iron-pickaxe")).isPresent());
+		assertTrue(registry.find(new OpenIdentifier("forgero:special_pickaxe_handle")).isPresent());
+	}
+}
